requirements.txt
absl-py==1.4.0
astunparse==1.6.3
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.2.0
cloudpickle==2.2.1
decorator==5.1.1
dm-reverb==0.12.0
dm-tree==0.1.8
flatbuffers==23.5.26
gast==0.4.0
gin-config==0.5.0
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.56.0
gym==0.23.0
gym-notices==0.0.8
h5py==3.9.0
idna==3.4
importlib-metadata==6.8.0
keras==2.13.1
libclang==16.0.0
lugo4py==0.0.3
Markdown==3.4.3
MarkupSafe==2.1.3
numpy==1.24.3
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.1
Pillow==10.0.0
portpicker==1.5.2
protobuf==4.23.4
psutil==5.9.5
pyasn1==0.5.0
pyasn1-modules==0.3.0
pygame==2.1.3
requests==2.31.0
requests-oauthlib==1.3.1
rlds==0.1.8
rsa==4.9
six==1.16.0
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.32.0
tensorflow-probability==0.20.1
termcolor==2.3.0
tf-agents==0.17.0
typing_extensions==4.5.0
urllib3==1.26.16
Werkzeug==2.3.6
wrapt==1.15.0
zipp==3.16.0
